backend/requirements.txt
# 项目基础依赖
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
# SQLite数据库依赖
python-dotenv>=1.0.1
pydantic>=2.10.6
openai>=1.66.3
markdown==3.5.1
python-multipart==0.0.6
gitpython==3.1.40
toml==0.10.2
tiktoken>=0.5.1

# OpenManus依赖
tenacity~=9.0.0
pyyaml~=6.0.2
loguru~=0.7.3
numpy
datasets~=3.2.0
html2text~=2024.2.26
gymnasium~=1.0.0
pillow~=10.4.0
browsergym~=0.13.3
unidiff~=0.7.5
browser-use~=0.1.40
googlesearch-python~=1.3.0
baidusearch~=1.0.3
duckduckgo_search~=7.5.1
aiofiles~=24.1.0
pydantic_core~=2.27.2
colorama~=0.4.6
playwright~=1.50.0
docker~=7.1.0
pytest~=8.3.5
pytest-asyncio~=0.25.3